+++ packages/cli/package.json
@@ -28,7 +28,7 @@
     "build:types": "tsc -p tsconfig.build.json",
     "build:types:watch": "yarn run build:types --watch",
     "build:refdocs": "ts-node docsgen docs/cli.md",
-    "write-git-data": "node lib/gitData/scripts/getGitData.js",
+    "write-git-data": "node lib/util/gitData/writeGitData.js",
     "check-types": "tsc",
     "lint": "eslint --color --ext .ts src/ test/",
     "lint:fix": "yarn run lint --fix",
