+++ packages/cli/src/util/gitData/index.ts
@@ -1,10 +1,9 @@
-import fs from "fs";
 import {execSync} from "child_process";
-import {getLocalVersion} from "./version";
+import {getLocalVersion} from "../version";
 
 // This file is created in the build step and is distributed through NPM
-// MUST be in sync with packages/cli/scripts/getGitData.js, and package.json .files
-import {gitDataPath, GitDataFile} from "../gitData/gitDataPath";
+// MUST be in sync with packages/cli/src/gitData/gitDataPath.ts, and package.json .files
+import {GitDataFile, readGitDataFile} from "./gitDataPath";
 
 /* eslint-disable @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-assignment */
 type GitData = {
@@ -18,16 +17,38 @@ type GitData = {
   version: string;
 };
 
+/** Silent shell that won't pollute stdout, or stderr */
+function shell(cmd: string): string {
+  return execSync(cmd, {stdio: ["ignore", "pipe", "ignore"]})
+    .toString()
+    .trim();
+}
+
+function getBranch(): string | undefined {
+  try {
+    return shell("git rev-parse --abbrev-ref HEAD");
+  } catch (e) {
+    return undefined;
+  }
+}
+
+function getCommit(): string | undefined {
+  try {
+    return shell("git rev-parse --verify HEAD");
+  } catch (e) {
+    return undefined;
+  }
+}
+
+export function getGitData(): Partial<Pick<GitData, "branch" | "commit">> {
+  return {
+    branch: getBranch(),
+    commit: getCommit(),
+  };
+}
+
 /**
- * Reads git data injected on the Dockerfile to the filepath `DOCKER_LODESTAR_GIT_DATA_FILEPATH`
- * Expects a file with JSON contents:
- * ```js
- * {
- *   version: "0.16.0",
- *   branch: "developer/feature-1",
- *   commit: "4f816b16dfde718e2d74f95f2c8292596138c248"
- * }
- * ```
+ * Reads git data from a persisted file at build time + the current version in the package.json
  */
 export function readLodestarGitData(): GitData {
   try {
@@ -52,32 +73,9 @@ function formatVersion({semver, branch, commit}: Partial<GitData>): string {
   return [semver, branch, commit && commit.slice(0, 8)].filter((s) => s).join(" ");
 }
 
-function getGitData(): Partial<GitData> {
+function getPersistedGitData(): Partial<GitDataFile> {
   try {
-    const shell = (cmd: string): string =>
-      execSync(cmd, {stdio: ["ignore", "pipe", "ignore"]})
-        .toString()
-        .trim();
-    const branch = shell("git rev-parse --abbrev-ref HEAD");
-    const commit = shell("git rev-parse --verify HEAD");
-    return {branch, commit};
-  } catch (e) {
-    return {};
-  }
-}
-
-function getPersistedGitData(): Partial<GitData> {
-  try {
-    const gitData = JSON.parse(fs.readFileSync(gitDataPath, "utf8")) as GitDataFile;
-    const {semver, branch, commit} = gitData;
-
-    const version = [semver, branch, commit && commit.slice(0, 8)].filter((s) => s).join(" ");
-    return {
-      semver: semver || "-",
-      branch: branch || "-",
-      commit: commit || "-",
-      version: version || "-",
-    };
+    return readGitDataFile();
   } catch (e) {
     return {};
   }
