+++ packages/cli/src/util/index.ts
@@ -4,6 +4,7 @@ export * from "./ethers";
 export * from "./file";
 export * from "./format";
 export * from "./fs";
+export * from "./gitData";
 export * from "./graffiti";
 export * from "./logger";
 export * from "./object";
